Add explosion when hit by enemy
When player is hit by the enemy, the explosion 
sprite is moved to the player position and the 
animation is played.

Issues:  The sprite starts in the middle of the screen
on the last frame, which isn't perfectly transparent, 
so leaves a small cloud.  Also, when the explosion 
is activated, it leaves behind the last frame of the 
cloud.  To improve this, the sprite should be 
hidden or destroyed after the animation is played.